pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "pydoftools"
version = "0.0.1"
authors = [
    { name="zhyipeng", email="zhyipeng@outlook.com" },
]
description = "A python library for dof-taiwan"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "anyio>=3.6.2",
    "loguru>=0.7.0",
    "Pillow>=9.5.0",
    "zhconv>=1.4.3",
]

[project.optional-dependencies]
tqdm = ["tqdm"]
